# Improved performance of unary _not_ for aligned bitmaps (3x)