```
bitmap aligned not 2^20 time:   [32.273 us 32.387 us 32.524 us]                                     
                        change: [-60.815% -60.557% -60.274%] (p = 0.00 < 0.05)
bitmap not 2^20         time:   [76.473 us 76.545 us 76.630 us]                            
                        change: [-8.2910% -8.1104% -7.9196%] (p = 0.00 < 0.05)
```